requirements.txt
opencv-python
numpy
pyserial
pynput==1.6.8
pyinstaller
